refactor: 경고 발생 수정 시도
