+++ week1/practice-oop-lotto/src/main/java/inputController/WinningNumbersInputController.java
@@ -1,7 +1,7 @@
 package inputController;
 
 import global.MessageManager;
-import viewInterface.InputController;
+import viewController.InputController;
 
 public class WinningNumbersInputController extends InputController {
     @Override
