+++ week1/practice-oop-lotto/src/main/java/viewController/OutputController.java
@@ -1,4 +1,4 @@
-package viewInterface;
+package viewController;
 
 import model.MModel;
