+++ week1/practice-oop-lotto/src/main/java/viewController/ViewController.java
@@ -1,4 +1,4 @@
-package viewInterface;
+package viewController;
 
 public interface ViewController {
     String getMessage();
